Turn the minigallery and generate directives into extensions
