+++ sunpy/util/sphinx/generate.py
@@ -2,11 +2,10 @@ import os
 import sys
 from io import StringIO
 
-from docutils import nodes, statemachine
-from docutils.io import FileInput
+from docutils import nodes
 from docutils.parsers.rst import Directive, directives
 
-__all__ = ['MiniGallery', 'Generate']
+__all__ = ['Generate']
 
 
 class Generate(Directive):
@@ -57,40 +56,5 @@ class Generate(Directive):
             sys.stdout = old_stdout
 
 
-class MiniGallery(Directive):
-    """
-    Custom directive to insert a mini-gallery
-
-    The required argument is the qualified name of the object.  The mini-gallery will be the subset
-    of gallery examples that make use of that object (from that specific namespace).  There can be
-    more than one object named, separated by spaces.
-    """
-    required_arguments = 1
-    optional_arguments = 0
-    final_argument_whitespace = True
-
-    def run(self):
-        # Respect the same disabling options as the ``raw`` directive
-        if (not self.state.document.settings.raw_enabled
-                or not self.state.document.settings.file_insertion_enabled):
-            raise self.warning('"%s" directive disabled.' % self.name)
-
-        obj_list = self.arguments[0].split()
-
-        # Concatenate the backreferences file(s)
-        lines = []
-        for obj in obj_list:
-            path = os.path.join(os.getcwd(), 'generated', 'modules', f'{obj}.examples')
-            lines += (FileInput(source_path=path).readlines())[5:]  # slice removes heading
-
-        # Append the end for the gallery
-        lines += ['\n',
-                  '.. raw:: html\n',
-                  '\n',
-                  '    <div class="sphx-glr-clear"></div>\n']
-        text = ''.join(lines)
-
-        include_lines = statemachine.string2lines(text, convert_whitespace=True)
-        self.state_machine.insert_input(include_lines, path)
-
-        return []
+def setup(app):
+    app.add_directive('generate', Generate)
